Renamed shallow and deep workers to fast and hard workers.
Hard workers are now the norm. When the sim starts running, if there's a fast worker, it renders at regular speed to the screen, until the hard worker renders.

The hard worker needs to alert the main thread that it's running slow in order for the fast worker to render in its place.